backend_v2/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-ninja==0.17.0
django-ninja-extra==0.14.8
django-oauth-toolkit==1.7.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
inflection==0.5.1
injector==0.19.0
itypes==1.2.0
Jinja2==3.1.1
jwcrypto==1.0
MarkupSafe==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.0
